Display machine deployment statuses
